Updated readme with todo list, added comments and more pocketsphinx recognition code

+++ src/emotion_game/include/MimicGame.h
@@ -33,7 +33,7 @@ private:
 	int _userToTrack;
 	ros::Subscriber _classSubscriber;
 	std::vector<nao_autism_messages::PoseClassification> _poseQueue;
-	short _currentPoseClassification;
+	int _currentPoseClassification;
 
 	void setOverallClassification(void);
 	void classificationCallback(const nao_autism_messages::PoseClassification poseClass);
